requirements.txt
PyGithub
Selenium